+++ src/extensions/mvs/additions/annotation-label/visual.ts
@@ -16,8 +16,8 @@ import { Theme } from '../../../../mol-theme/theme';
 import { ColorNames } from '../../../../mol-util/color/names';
 import { ParamDefinition as PD } from '../../../../mol-util/param-definition';
 import { textPropsForSelection } from '../../helpers/label-text';
-import { AnnotationRow } from '../../helpers/schemas';
-import { omitObjectKeys, range } from '../../helpers/utils';
+import { groupRows } from '../../helpers/selections';
+import { omitObjectKeys } from '../../helpers/utils';
 import { getAnnotationForStructure } from '../annotation-prop';
 
 
@@ -63,54 +63,3 @@ function createLabelText(ctx: VisualContext, structure: Structure, theme: Theme,
     }
     return builder.getText();
 }
-
-/** Data structure for an array divided into contiguous groups */
-interface GroupedArray<T> {
-    /** Number of groups */
-    count: number,
-    /** Get size of i-th group as `offsets[i+1]-offsets[i]`.
-     * Get j-th element in i-th group as `grouped[offsets[i]+j]` */
-    offsets: number[],
-    /** Get j-th element in i-th group as `grouped[offsets[i]+j]` */
-    grouped: T[],
-}
-
-/** Return row indices grouped by `row.group_id`. Rows with `row.group_id===undefined` are treated as separate groups. */
-function groupRows(rows: readonly AnnotationRow[]): GroupedArray<number> {
-    let counter = 0;
-    const groupMap = new Map<string, number>();
-    const groups: number[] = [];
-    for (let i = 0; i < rows.length; i++) {
-        const group_id = rows[i].group_id;
-        if (group_id === undefined) {
-            groups.push(counter++);
-        } else {
-            const groupIndex = groupMap.get(group_id);
-            if (groupIndex === undefined) {
-                groupMap.set(group_id, counter);
-                groups.push(counter);
-                counter++;
-            } else {
-                groups.push(groupIndex);
-            }
-        }
-    }
-    const rowIndices = range(rows.length).sort((i, j) => groups[i] - groups[j]);
-    const offsets: number[] = [];
-    for (let i = 0; i < rows.length; i++) {
-        if (i === 0 || groups[rowIndices[i]] !== groups[rowIndices[i - 1]]) offsets.push(i);
-    }
-    offsets.push(rowIndices.length);
-    return { count: offsets.length - 1, offsets, grouped: rowIndices };
-}
-export function testGroupRows() {
-    const rows = [{ label: 'A' }, { label: 'B', group_id: 1 }, { label: 'C', group_id: 'x' }, { label: 'D', group_id: 1 }, { label: 'E' }, { label: 'F' }, { label: 'G', group_id: 'x' }, { label: 'H', group_id: 'x' }] as any as AnnotationRow[];
-    const { count, offsets, grouped } = groupRows(rows);
-    for (let i = 0; i < count; i++) {
-        console.log('Group', i);
-        for (let j = offsets[i], stop = offsets[i + 1]; j < stop; j++) {
-            console.log('   ', rows[grouped[j]]);
-        }
-    }
-}
-// TODO turn into a proper jest test
