+++ src/extensions/mvs/helpers/selections.ts
@@ -317,3 +317,44 @@ export function rowsToExpression(rows: readonly AnnotationRow[]): Expression {
 function unionExpression(expressions: Expression[]): Expression {
     return MS.struct.combinator.merge(expressions.map(e => MS.struct.modifier.union([e])));
 }
+
+
+/** Data structure for an array divided into contiguous groups */
+interface GroupedArray<T> {
+    /** Number of groups */
+    count: number,
+    /** Get size of i-th group as `offsets[i+1]-offsets[i]`.
+     * Get j-th element in i-th group as `grouped[offsets[i]+j]` */
+    offsets: number[],
+    /** Get j-th element in i-th group as `grouped[offsets[i]+j]` */
+    grouped: T[],
+}
+
+/** Return row indices grouped by `row.group_id`. Rows with `row.group_id===undefined` are treated as separate groups. */
+export function groupRows(rows: readonly AnnotationRow[]): GroupedArray<number> {
+    let counter = 0;
+    const groupMap = new Map<string, number>();
+    const groups: number[] = [];
+    for (let i = 0; i < rows.length; i++) {
+        const group_id = rows[i].group_id;
+        if (group_id === undefined) {
+            groups.push(counter++);
+        } else {
+            const groupIndex = groupMap.get(group_id);
+            if (groupIndex === undefined) {
+                groupMap.set(group_id, counter);
+                groups.push(counter);
+                counter++;
+            } else {
+                groups.push(groupIndex);
+            }
+        }
+    }
+    const rowIndices = range(rows.length).sort((i, j) => groups[i] - groups[j]);
+    const offsets: number[] = [];
+    for (let i = 0; i < rows.length; i++) {
+        if (i === 0 || groups[rowIndices[i]] !== groups[rowIndices[i - 1]]) offsets.push(i);
+    }
+    offsets.push(rowIndices.length);
+    return { count: offsets.length - 1, offsets, grouped: rowIndices };
+}
