+++ src/extensions/mvs/load.ts
@@ -26,8 +26,7 @@ import { MVSTreeSchema } from './tree/mvs/mvs-tree';
 
 
 /** Load a MolViewSpec (MVS) tree into the Mol* plugin.
- * If `deletePrevious`, remove all objects in the current Mol* state; otherwise add to the current state.
- */
+ * If `deletePrevious`, remove all objects in the current Mol* state; otherwise add to the current state. */
 export async function loadMVS(plugin: PluginContext, data: MVSData, deletePrevious: boolean) {
     // console.log(`MVS tree (v${data.version}):\n${treeToString(data.root)}`);
     validateTree(MVSTreeSchema, data.root, 'MVS');
@@ -39,11 +38,10 @@ export async function loadMVS(plugin: PluginContext, data: MVSData, deletePrevio
 
 
 /** Load a `MolstarTree` into the Mol* plugin.
- * If `deletePrevious`, remove all objects in the current Mol* state; otherwise add to the current state.
- */
+ * If `deletePrevious`, remove all objects in the current Mol* state; otherwise add to the current state. */
 async function loadMolstarTree(plugin: PluginContext, tree: MolstarTree, deletePrevious: boolean) {
     const mvsExtensionLoaded = plugin.spec.behaviors.some(b => b.transformer.id === MolViewSpec.id);
-    if (!mvsExtensionLoaded) { 
+    if (!mvsExtensionLoaded) {
         console.warn('MolViewSpec extension is not loaded.');
         throw new Error('MolViewSpec extension is not loaded.');
     }
